Remove Travis and Cirrus, use GH also for FreeBSD (#1880)

+++ scripts/internal/download_wheels_github.py
@@ -57,12 +57,14 @@ def rename_win27_wheels():
     # See: https://github.com/giampaolo/psutil/issues/810
     src = 'dist/psutil-%s-cp27-cp27m-win32.whl' % PSUTIL_VERSION
     dst = 'dist/psutil-%s-cp27-none-win32.whl' % PSUTIL_VERSION
-    print("rename: %s\n        %s" % (src, dst))
-    os.rename(src, dst)
+    if os.path.exists(src):
+        print("rename: %s\n        %s" % (src, dst))
+        os.rename(src, dst)
     src = 'dist/psutil-%s-cp27-cp27m-win_amd64.whl' % PSUTIL_VERSION
     dst = 'dist/psutil-%s-cp27-none-win_amd64.whl' % PSUTIL_VERSION
-    print("rename: %s\n        %s" % (src, dst))
-    os.rename(src, dst)
+    if os.path.exists(src):
+        print("rename: %s\n        %s" % (src, dst))
+        os.rename(src, dst)
 
 
 def run():
